Refactor: Rename post_create.ts to post_employee.ts for consistency
