+++ api/src/routes/employees/index.ts
@@ -4,7 +4,7 @@ import { dbMiddleware } from '../../middleware';
 
 import getListApp from './get_list';
 import getDetailApp from './get_detail';
-import postCreateApp from './post_create';
+import postEmployeeApp from './post_employee';
 import putUpdateApp from './put_update';
 import deleteDeleteApp from './delete_delete';
 
@@ -17,7 +17,7 @@ app.use('*', dbMiddleware());
 // 各APIをマウント
 app.route('/', getListApp);             // 一覧取得
 app.route('/:id', getDetailApp);        // 詳細取得
-app.route('/', postCreateApp);          // 新規作成
+app.route('/', postEmployeeApp);        // 新規作成
 app.route('/:id', putUpdateApp);        // 更新
 app.route('/:id', deleteDeleteApp);     // 削除
